refactor: major isolation refactor for clarity
