+++ core/elements/src/isolation/boot/import-map.ts
@@ -1,0 +1,56 @@
+/**
+ * Reads the host page's `<script type="importmap">` and resolves every URL in
+ * it to absolute (against the host's `baseURI`), so the same import map can be
+ * handed to the iframe's es-module-shims where relative bases don't apply.
+ */
+
+export interface ImportMap {
+  imports?: Record<string, string>;
+  scopes?: Record<string, Record<string, string>>;
+}
+
+/** Read the host page's import map and resolve all URLs to absolute. */
+export function getResolvedImportMap(): ImportMap {
+  const script = document.querySelector('script[type="importmap"]');
+  if (!script?.textContent) return {};
+  try {
+    const raw: ImportMap = JSON.parse(script.textContent);
+    const baseURI = document.baseURI;
+    const resolved: ImportMap = {};
+
+    if (raw.imports) {
+      resolved.imports = {};
+      for (const [key, value] of Object.entries(raw.imports)) {
+        try {
+          resolved.imports[key] = new URL(value, baseURI).href;
+        } catch {
+          resolved.imports[key] = value;
+        }
+      }
+    }
+
+    if (raw.scopes) {
+      resolved.scopes = {};
+      for (const [scopeKey, scopeMap] of Object.entries(raw.scopes)) {
+        let resolvedKey: string;
+        try {
+          resolvedKey = new URL(scopeKey, baseURI).href;
+        } catch {
+          resolvedKey = scopeKey;
+        }
+        resolved.scopes[resolvedKey] = {};
+        for (const [k, v] of Object.entries(scopeMap)) {
+          try {
+            resolved.scopes[resolvedKey][k] = new URL(v, baseURI).href;
+          } catch {
+            resolved.scopes[resolvedKey][k] = v;
+          }
+        }
+      }
+    }
+
+    return resolved;
+  } catch {
+    return {};
+  }
+}
