+++ core/elements/src/isolation/log.ts
@@ -1,0 +1,4 @@
+import debug from "debug";
+
+/** Shared debug logger for the isolation boundary (host side). */
+export const log = debug("patchwork:elements:isolation");
